Revert "Add ancillaries fast track  (#1577)"
This reverts commit d1f40c5b836329c8df159c935074d8622e917968.
