--- app/ancillaryFactory/index.js
@@ -1,5 +1,0 @@
-// @flow strict
-
-export {
-  default as AncillaryFactory,
-} from './src/appRegistry/AncillaryFactory';
