--- app/ancillaryFactory/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@kiwicom/react-native-ancillary-factory",
-  "version": "0.0.1",
-  "private": true,
-  "main": "index.js",
-  "dependencies": {
-    "@kiwicom/mobile-config": "^0",
-    "@kiwicom/mobile-localization": "^0",
-    "@kiwicom/react-native-fast-track": "^0",
-    "@kiwicom/mobile-shared": "^0",
-    "@kiwicom/mobile-relay": "^0"
-  },
-  "devDependencies": {
-    "@kiwicom/mobile-playground": "^0"
-  },
-  "peerDependencies": {
-    "react": ">=16",
-    "react-native": ">=0.57",
-    "react-test-renderer": "^16",
-    "react-navigation": "^3",
-    "react-native-maps": "^0",
-    "react-navigation-props-mapper": "^0",
-    "@kiwicom/universal-components": "^0"
-  }
-}
