--- app/ancillaryFactory/src/config/ancillariesDefinitions.js
@@ -1,15 +1,0 @@
-// @flow
-
-import { FastTrackBanner } from '@kiwicom/react-native-fast-track';
-
-export type AncillaryDefinition = {
-  service: string,
-  renderComponent: Function,
-};
-
-export const ancillariesDefinitions = Object.freeze({
-  FAST_TRACK: {
-    service: 'fast_track',
-    renderComponent: FastTrackBanner,
-  },
-});
