+++ package.json
@@ -30,10 +30,9 @@
   "dependencies": {
     "@kiwicom/mobile-shared": "^0",
     "@kiwicom/react-native-app-hotels": "^0",
-    "@kiwicom/react-native-ancillary-factory": "^0",
     "@kiwicom/react-native-fast-track": "^0",
     "chalk": "^2.4.2",
-    "detox": "^12.1.3",
+    "detox": "^12.1.2",
     "dotenv": "^7.0.0",
     "react": "16.8.3",
     "react-native": "0.59.2",
