+++ packages/localization/src/vocabularies/FastTrack.js
@@ -1,15 +1,5 @@
 // @flow strict
 
-const FastTrackVocabulary = {
-  'fast_track.banner.title': 'Milan Bergamo Stopover Bundle',
-  'fast_track.banner.description':
-    'VIP treatment is waiting for you at Milan Bergamo Airport with Fast Track and baggage assistance.',
-  'fast_track.banner.services_label': 'Services included',
-  'fast_track.banner.services_security_route': 'Fast track security route',
-  'fast_track.banner.services_handling_assistance':
-    'Baggage handling assistance',
-  'fast_track.banner.show_qr_button': 'Show QR Code',
-  'fast_track.banner.modal.close_modal_button': 'Close modal',
-};
+const FastTrackVocabulary = {};
 
 export default FastTrackVocabulary;
